Refactoring Matts changes
Moving the LoginUX files to be linked from a higher level
