+++ cds/App.config
@@ -2,8 +2,11 @@
 <configuration>
   <connectionStrings>
 
-    <!--Un-comment and edit one of the example connection strings below using appropriate values for your environment -->
-
+    <!--
+    To run all samples without being prompted for login credentials, un-comment and edit one of the example connection strings below using appropriate values for your environment.
+    Otherwise, you will be prompted for login information when you run a sample project.   
+    -->
+    
     <!-- 
       ** !IMPORTANT! **
 
@@ -15,19 +18,6 @@
         information you use to run the samples. If you push your changes to this file, 
         your connection information including your login and password information will 
         be visible to anyone with access to your cloned repository.
-
-        Before you edit this file, run the following GIT command to remove this file 
-        from the index of tracked files for your repo:
-
-
-        > git update-index <hyphen><hyphen>assume-unchanged cds/App.config
-
-        Note: Replace <hyphen> with '-'. Double hyphen characters are not valid within an XML comment.
-
-
-
-        Git documentation: https://git-scm.com/docs/git-update-index
-
     -->
 
     <!--Online using Office 365-->
@@ -80,14 +70,35 @@ connectionString="Url=https://yourOrgName.yourServer.com/yourOrgName; Username=y
 								Warning:    includes Error, Warning levels
 								Error:      includes Error level
 				-->
-      <add name="Microsoft.Xrm.Tooling.Connector.CrmServiceClient" value="Verbose"/>
-      <add name="Microsoft.Xrm.Tooling.CrmConnectControl" value="Verbose"/>
+      <add name="Microsoft.Xrm.Tooling.Connector.CrmServiceClient" value="Error"/>
+      <add name="Microsoft.Xrm.Tooling.CrmConnectControl" value="Error"/>
       <add name="Microsoft.IdentityModel.Clients.ActiveDirectory" value="Error"/>
     </switches>
     <sharedListeners>
-      <add name="fileListener" type="Microsoft.Xrm.Tooling.Connector.DynamicsFileLogTraceListener, Microsoft.Xrm.Tooling.Connector" BaseFileName="CRUD-Dynamic-Entity" Location="LocalUserApplicationDirectory" MaxFileSize="52428800"/>
-      <add name="ADALListener" type="Microsoft.Xrm.Tooling.Connector.DynamicsFileLogTraceListener, Microsoft.Xrm.Tooling.Connector" BaseFileName="CRUD-Dynamic-Entity-ADAL" Location="LocalUserApplicationDirectory" MaxFileSize="52428800"/>
+      <add name="fileListener"
+  type="Microsoft.Xrm.Tooling.Connector.DynamicsFileLogTraceListener, Microsoft.Xrm.Tooling.Connector" />
+      <add name="fileListener"
+        type="Microsoft.Xrm.Tooling.Connector.DynamicsFileLogTraceListener, Microsoft.Xrm.Tooling.Connector"
+        BaseFileName="PowerApps-Sample-Log"
+        Location="LocalUserApplicationDirectory"
+        MaxFileSize="52428800"/>
+      <!--
+      Log written to 
+      \\{Computer Name}\Users\{UserName}\AppData\Roaming\PowerApps\{Sample Project Name}\1.0.0.0\PowerApps-Sample-Log-{YYYY}-{MM}-{DD}.log
+      -->
+      <add name="ADALListener"
+        type="Microsoft.Xrm.Tooling.Connector.DynamicsFileLogTraceListener, Microsoft.Xrm.Tooling.Connector"
+        BaseFileName="PowerApps-Sample-Log-ADAL"
+        Location="LocalUserApplicationDirectory"
+        MaxFileSize="52428800"/>
+      <!--
+      Log written to 
+      \\{Computer Name}\Users\{UserName}\AppData\Roaming\PowerApps\{Sample Project Name}\1.0.0.0\PowerApps-Sample-Log-ADAL-{YYYY}-{MM}-{DD}.log
+      -->
     </sharedListeners>
   </system.diagnostics>
 
-<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6.2"/></startup></configuration>
+  <startup>
+    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6.2"/>
+  </startup>
+</configuration>
