+++ cds/orgsvc/C#/CRUD-Dynamic-Entity/CRUD-Dynamic-Entity/CRUD-Dynamic-Entity.csproj
@@ -108,11 +108,11 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\..\..\..\LoginUX\ExampleLoginForm.xaml.cs">
+      <Link>LoginUX\ExampleLoginForm.xaml.cs</Link>
+    </Compile>
     <Compile Include="..\..\SampleHelpers.cs">
       <Link>SampleHelpers.cs</Link>
-    </Compile>
-    <Compile Include="LoginUX\ExampleLoginForm.xaml.cs">
-      <DependentUpon>ExampleLoginForm.xaml</DependentUpon>
     </Compile>
     <Compile Include="SampleProgram.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -123,11 +123,11 @@
     </None>
     <None Include="packages.config" />
   </ItemGroup>
-  <ItemGroup />
   <ItemGroup>
-    <Page Include="LoginUX\ExampleLoginForm.xaml">
-      <SubType>Designer</SubType>
+    <Page Include="..\..\..\..\LoginUX\ExampleLoginForm.xaml">
+      <Link>LoginUX\ExampleLoginForm.xaml</Link>
       <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
     </Page>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
