+++ cds/orgsvc/C#/SampleHelpers.cs
@@ -1,10 +1,9 @@
-﻿using Microsoft.Crm.Sdk.Messages;
-using Microsoft.Xrm.Sdk;
+﻿using Microsoft.Xrm.Sdk;
 using Microsoft.Xrm.Sdk.Query;
 using Microsoft.Xrm.Tooling.Connector;
+using PowerApps.Samples.LoginUX;
 using System;
 using System.Configuration;
-using System.IO;
 using System.ServiceModel;
 
 namespace PowerApps.Samples
@@ -19,10 +18,6 @@ namespace PowerApps.Samples
         /// <returns>true when the version is higher than the minimum verions, otherwise false.</returns>
         public static bool CheckVersion(CrmServiceClient service, Version minVersion)
         {
-            //RetrieveVersionResponse crmVersionResp = (RetrieveVersionResponse)service.Execute(new RetrieveVersionRequest());
-
-            //Version currentVersion = new Version(crmVersionResp.Version);
-
             if (service.ConnectedOrgVersion.CompareTo(minVersion) >= 0)
             {
                 return true;
@@ -61,16 +56,7 @@ namespace PowerApps.Samples
             {
                 Console.WriteLine("The {0} solution is not installed. Importing the solution....", uniqueName);
                 Guid ImportId = Guid.Empty;
-                service.ImportSolutionToCrm(pathToFile, out ImportId); 
-                
-                //byte[] fileBytes = File.ReadAllBytes(pathToFile);
-                //ImportSolutionRequest impSolReq = new ImportSolutionRequest()
-                //{
-                //    CustomizationFile = fileBytes, 
-                //};
-
-                //service.Execute(impSolReq);
-
+                service.ImportSolutionToCrm(pathToFile, out ImportId);                
                 return true;
             }
         }
@@ -101,7 +87,7 @@ namespace PowerApps.Samples
                 solutionQuery.Criteria.AddCondition("uniquename", ConditionOperator.Equal, uniqueName);
 
 
-                Entity solution = (Entity)service.RetrieveMultiple(solutionQuery).Entities[0];
+                Entity solution = service.RetrieveMultiple(solutionQuery).Entities[0];
 
                 if (solution != null)
                 {
@@ -120,51 +106,50 @@ namespace PowerApps.Samples
         /// <summary>
         /// A function to manage exceptions thrown by console application samples
         /// </summary>
-        /// <param name="ex">The exception thrown</param>
-        public static void HandleException(Exception ex) {
+        /// <param name="exceptionFromSample">The exception thrown</param>
+        public static void HandleException(Exception exceptionFromSample) {
             Console.WriteLine("The application terminated with an error.");
 
-            switch (ex) {
-
-                case FaultException<OrganizationServiceFault> fe:
-
-                    Console.WriteLine("Timestamp: {0}", fe.Detail.Timestamp);
-                    Console.WriteLine("Code: {0}", fe.Detail.ErrorCode);
-                    Console.WriteLine("Message: {0}", fe.Detail.Message);
-                    Console.WriteLine("Plugin Trace: {0}", fe.Detail.TraceText);
-                    Console.WriteLine("Inner Fault: {0}",
-                        null == fe.Detail.InnerFault ? "No Inner Fault" : "Has Inner Fault");
-                    break;
-                case TimeoutException te:
-                    
-                    Console.WriteLine("Message: {0}", te.Message);
-                    Console.WriteLine("Stack Trace: {0}", te.StackTrace);
-                    Console.WriteLine("Inner Fault: {0}",
-                        null == te.InnerException.Message ? "No Inner Fault" : te.InnerException.Message);
-                    break;
-                // Additional exceptions to catch: SecurityTokenValidationException, ExpiredSecurityTokenException,
-                // SecurityAccessDeniedException, MessageSecurityException, and SecurityNegotiationException.
-                default:
-
-                    // Display the details of the inner exception.
-                    if (ex.InnerException != null)
-                    {
-                        Console.WriteLine(ex.InnerException.Message);
-
-                        FaultException<OrganizationServiceFault> fe = ex.InnerException
-                            as FaultException<OrganizationServiceFault>;
-                        if (fe != null)
-                        {
-                            Console.WriteLine("Timestamp: {0}", fe.Detail.Timestamp);
-                            Console.WriteLine("Code: {0}", fe.Detail.ErrorCode);
-                            Console.WriteLine("Message: {0}", fe.Detail.Message);
-                            Console.WriteLine("Plugin Trace: {0}", fe.Detail.TraceText);
-                            Console.WriteLine("Inner Fault: {0}",
-                                null == fe.Detail.InnerFault ? "No Inner Fault" : "Has Inner Fault");
-                        }
-                    }
-                    break;
+            try
+            {
+                throw exceptionFromSample;
             }
+            catch (FaultException<OrganizationServiceFault> fe)
+            {
+                Console.WriteLine("Timestamp: {0}", fe.Detail.Timestamp);
+                Console.WriteLine("Code: {0}", fe.Detail.ErrorCode);
+                Console.WriteLine("Message: {0}", fe.Detail.Message);
+                Console.WriteLine("Plugin Trace: {0}", fe.Detail.TraceText);
+                Console.WriteLine("Inner Fault: {0}",
+                    null == fe.Detail.InnerFault ? "No Inner Fault" : "Has Inner Fault");
+            }
+            catch (TimeoutException te) {
+                Console.WriteLine("Message: {0}", te.Message);
+                Console.WriteLine("Stack Trace: {0}", te.StackTrace);
+                Console.WriteLine("Inner Fault: {0}",
+                    null == te.InnerException.Message ? "No Inner Fault" : te.InnerException.Message);
+
+            }
+            catch (Exception ex) {
+                // Display the details of the inner exception.
+                if (ex.InnerException != null)
+                {
+                    Console.WriteLine(ex.InnerException.Message);
+
+                    FaultException<OrganizationServiceFault> fe = ex.InnerException
+                        as FaultException<OrganizationServiceFault>;
+                    if (fe != null)
+                    {
+                        Console.WriteLine("Timestamp: {0}", fe.Detail.Timestamp);
+                        Console.WriteLine("Code: {0}", fe.Detail.ErrorCode);
+                        Console.WriteLine("Message: {0}", fe.Detail.Message);
+                        Console.WriteLine("Plugin Trace: {0}", fe.Detail.TraceText);
+                        Console.WriteLine("Inner Fault: {0}",
+                            null == fe.Detail.InnerFault ? "No Inner Fault" : "Has Inner Fault");
+                    }
+                }
+            }
+            
         }
 
         /// <summary>
@@ -172,7 +157,7 @@ namespace PowerApps.Samples
         /// </summary>
         /// <param name="name">The name of the connection string to return</param>
         /// <returns>The named connection string</returns>
-        public static string GetConnectionStringFromAppConfig(string name)
+        private static string GetConnectionStringFromAppConfig(string name)
         {
             //Verify cds/App.config contains a valid connection string with the name.
             try
@@ -181,8 +166,50 @@ namespace PowerApps.Samples
             }
             catch (Exception)
             {
-                Console.WriteLine("You must set connection data in cds/App.config before running this sample. - Switching to Interactive Mode");
+                Console.WriteLine("You can set connection data in cds/App.config before running this sample. - Switching to Interactive Mode");
                 return string.Empty;
+            }
+        }
+
+        public static CrmServiceClient Connect(string name) {
+            CrmServiceClient service = null;
+            //You can specify connection information in cds/App.config to run this sample without the login dialog
+            if (string.IsNullOrEmpty(GetConnectionStringFromAppConfig("Connect")))
+            {
+                // Failed to find a connection string... Show login Dialog. 
+                ExampleLoginForm loginFrm = new ExampleLoginForm();
+                // Login process is Async, thus we need to detect when login is completed and close the form. 
+                loginFrm.ConnectionToCrmCompleted += LoginFrm_ConnectionToCrmCompleted;
+                // Show the dialog here. 
+                loginFrm.ShowDialog();
+
+                // If the login process completed, assign the connected service to the CRMServiceClient var 
+                if (loginFrm.CrmConnectionMgr != null && loginFrm.CrmConnectionMgr.CrmSvc != null && loginFrm.CrmConnectionMgr.CrmSvc.IsReady)
+                    service = loginFrm.CrmConnectionMgr.CrmSvc;
+                    
+                
+            }
+            else
+            {
+                // Try to create via connection string. 
+                service = new CrmServiceClient(GetConnectionStringFromAppConfig("Connect"));
+                
+            }
+
+            return service;
+
+        }
+
+        /// <summary>
+        /// Handle closing the dialog when completed. 
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private static void LoginFrm_ConnectionToCrmCompleted(object sender, EventArgs e)
+        {
+            if (sender is ExampleLoginForm)
+            {
+                ((ExampleLoginForm)sender).Close();
             }
         }
     }
